% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optinterval.R
\name{optinterval}
\alias{optinterval}
\title{Optimum Piecewise Segmentation of event arrival data}
\usage{
optinterval(data, N, c, type = c("constant"), alpha = 0.05,
  verbose = FALSE)
}
\arguments{
\item{data}{A vector of time-to-event cell arrival times \code{nx1}}

\item{N}{A vector containing the number of photons per cell \code{nx1}}

\item{c}{Block penalty scalar value \code{c>0}}

\item{type}{A vector containing the block types to be conisdered in the model 
\itemize{
  \item "constant" for constant blocks \code{y=a},
  \item "linear" for linear function blocks \code{y=ax+b},
  \item "exponential" for exponential function blocks \code{y=ae^(bx)+c},
  \item "linear" for power function blocks \code{y=at^b+c}
}}

\item{alpha}{The significance level desired for model comparison}

\item{verbose}{Logical indicating whether you want algorithm progress output}
}
\value{
A BB object containing all segmentation information
}
\description{
This function runs the PELT algorithm to partition photon event data into 
piecewise models. It accepts time-to-event, time-to-spill, and binned data.
}
\examples{
#create some random data
set.seed(1)
equation1 <- expression(40)
equation2 <- expression(40+100*x)
equation3 <- expression(240)
x1 <- NHPois(time_length =3,equation = equation1)
x2 <- NHPois(time_length =2,equation = equation2)+x1[length(x1)]
x3 <- NHPois(time_length =2,equation = equation3)+x2[length(x2)]
x=c(x1,x2,x3)
N <- rep(1,length(x))#let all cells be of size 1

#run model
model <- optinterval(x, N,3,type=c("constant","linear"),pvalue=0.05,verbose=TRUE)

#plot model
plot(model,show=c("blocks","points"),xlim=c(0,x[length(x)]),ylim=c(0,400))
}
