% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BB.R
\name{plot.BB}
\alias{plot.BB}
\title{plot}
\usage{
\method{plot}{BB}(x, show = c("hist", "blocks"), binwidth = NULL,
  bins = NULL, ylim = NULL, xlim = NULL, xact = NULL,
  main = "Bayesian Blocks", xlab = "Time", ylab = "Intensity")
}
\arguments{
\item{x}{A BB (Bayesian Block) object}

\item{show}{A vector containing what should be plotted
\itemize{
  \item "hist" to plot a histogram of the data,
  \item "blocks" to plot the piecewise blocks,
  \item "points" to plot all points in the data
}}

\item{binwidth}{binwidth desired for the histogram}

\item{bins}{number of bins desired for the histogram}

\item{...}{Arguments to be passed to methods, such as graphical parameters(see \code{\link{par}}). Many methods will accept the following arguments:}
}
\description{
This function plots a BayesianBlocks object
}
